+++ src/mruby_integration/models/font.cpp
@@ -16,12 +16,12 @@ void
 setup_Font(mrb_state* mrb,
            mrb_value object,
            Font* font,
-           int base_size,
+           int size,
            int glyph_count,
            int glyph_padding,
            Texture2D* texture)
 { //, Rectangle **rectangles) {
-  ivar_attr_int(mrb, object, font->baseSize, base_size);
+  ivar_attr_int(mrb, object, font->baseSize, size);
   ivar_attr_int(mrb, object, font->glyphCount, glyph_count);
   ivar_attr_int(mrb, object, font->glyphPadding, glyph_padding);
   ivar_attr_texture2d(mrb, object, font->texture, texture);
@@ -32,8 +32,25 @@ setup_Font(mrb_state* mrb,
 auto
 mrb_Font_initialize(mrb_state* mrb, mrb_value self) -> mrb_value
 {
-  mrb_int base_size, glyph_count, glyph_padding;
-  mrb_get_args(mrb, "iii", &base_size, &glyph_count, &glyph_padding);
+  char* path;
+
+  // def initialize(path, size: 32, glyph_count: 95)
+  mrb_int kw_num = 2;
+  mrb_int kw_required = 0;
+  mrb_sym kw_names[] = { mrb_intern_lit(mrb, "size"),
+                         mrb_intern_lit(mrb, "glyph_count") };
+  mrb_value kw_values[kw_num];
+  mrb_kwargs kwargs = { kw_num, kw_required, kw_names, kw_values, nullptr };
+  mrb_get_args(mrb, "z:", &path, &kwargs);
+
+  if (mrb_undef_p(kw_values[0])) {
+    kw_values[0] = mrb_fixnum_value(32);
+  }
+  if (mrb_undef_p(kw_values[1])) {
+    kw_values[1] = mrb_fixnum_value(95);
+  }
+
+  int* font_chars{ nullptr };
 
   Font* font = static_cast<Font*> DATA_PTR(self);
   if (font) {
@@ -42,33 +59,22 @@ mrb_Font_initialize(mrb_state* mrb, mrb_value self) -> mrb_value
   mrb_data_init(self, nullptr, &Font_type);
   font = static_cast<Font*>(malloc(sizeof(Font)));
 
-  Texture2D* texture = &font->texture;
+  *font = LoadFontEx(
+    path, mrb_fixnum(kw_values[0]), font_chars, mrb_fixnum(kw_values[1]));
 
-  setup_Font(mrb, self, font, base_size, glyph_count, glyph_padding, texture);
+  setup_Font(mrb,
+             self,
+             font,
+             font->baseSize,
+             font->glyphCount,
+             font->glyphPadding,
+             &font->texture);
 
   add_parent(font, "Font");
   add_owned_object(&font->texture);
 
   mrb_data_init(self, font, &Font_type);
   return self;
-}
-
-auto
-mrb_Font_set_base_size(mrb_state* mrb, mrb_value self) -> mrb_value
-{
-  attr_setter_int(mrb, self, Font_type, Font, baseSize, base_size);
-}
-
-auto
-mrb_Font_set_glyph_count(mrb_state* mrb, mrb_value self) -> mrb_value
-{
-  attr_setter_int(mrb, self, Font_type, Font, glyphCount, glyph_count);
-}
-
-auto
-mrb_Font_set_glyph_padding(mrb_state* mrb, mrb_value self) -> mrb_value
-{
-  attr_setter_int(mrb, self, Font_type, Font, glyphPadding, glyph_padding);
 }
 
 void
@@ -78,15 +84,6 @@ append_models_Font(mrb_state* mrb)
   MRB_SET_INSTANCE_TT(Font_class, MRB_TT_DATA);
   mrb_define_method(
     mrb, Font_class, "initialize", mrb_Font_initialize, MRB_ARGS_REQ(5));
-  mrb_define_method(
-    mrb, Font_class, "base_size=", mrb_Font_set_base_size, MRB_ARGS_REQ(1));
-  mrb_define_method(
-    mrb, Font_class, "glyph_count=", mrb_Font_set_glyph_count, MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    Font_class,
-                    "glyph_padding=",
-                    mrb_Font_set_glyph_padding,
-                    MRB_ARGS_REQ(1));
 
   load_ruby_models_font(mrb);
 }