+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Remove SimpleHTTPServer
 - Slight refactor of build pipeline
 - Group mruby gems in a gembox
+- Font#new now uses kwargs and is C++ backed
 
 ## v0.3.14.1